Life expectancy data

This function downloads the data for total life expectancy, male life expectancy and female life expectancy

Usage

gdt_wb_lex(country)

Arguments

country

A string with the country name or iso code

Value

A tibble with the requested data:

lex

Total life expectancy

lex_male

Male life expectancy

lex_female

Female life expectancy

Examples

gdt_wb_lex("Switzerland")
#> # A tibble: 64 × 7
#>    iso2c iso3c country      date lex_female   lex lex_male
#>    <chr> <chr> <chr>       <dbl>      <dbl> <dbl>    <dbl>
#>  1 CH    CHE   Switzerland  1960       74.1  71.3     68.6
#>  2 CH    CHE   Switzerland  1961       74.6  71.6     68.8
#>  3 CH    CHE   Switzerland  1962       74.1  71.2     68.4
#>  4 CH    CHE   Switzerland  1963       74.2  71.2     68.3
#>  5 CH    CHE   Switzerland  1964       75.1  72.1     69.2
#>  6 CH    CHE   Switzerland  1965       75.1  72.2     69.5
#>  7 CH    CHE   Switzerland  1966       75.2  72.3     69.6
#>  8 CH    CHE   Switzerland  1967       75.7  72.6     69.7
#>  9 CH    CHE   Switzerland  1968       75.5  72.6     69.8
#> 10 CH    CHE   Switzerland  1969       75.6  72.6     69.7
#> # ℹ 54 more rows