Soil elements data for Africa

Connect to or download chemical soil element concentration (for the 0-30 cm topsoil) data for Africa. The spatial resolution is 30 arc-seconds (about 1 km2), aggregated from the original 250 m spatial resolution.

The data have an Open Data Commons Open Database License (ODbL)

For more information, see https://web.archive.org/web/20250421195225/https://www.isric.org/projects/soil-property-maps-africa-250-m-resolution

Usage

soil_af_elements(var, path, ...)

Arguments

var

character. Variables name. One of: "Al", "B", "Ca", "Cu", "Fe", "K", "Mg", "Mn", "N", "Na", "P", "Ptot", "Zn". See Details

path

character. Path for storing the downloaded data. See geodata_path

...

additional arguments passed to download.file

Value

SpatRaster

Details

var	description	unit

Al	Extractable aluminum	mg kg-1 (ppm)

B	Extractable boron	mg kg-1 (ppm)

Ca	Extractable calcium	mg kg-1 (ppm)

Cu	Extractable copper	mg kg-1 (ppm)

Fe	Extractable iron	mg kg-1 (ppm)

K	Extractable potassium	mg kg-1 (ppm)

Mg	Extractable magnesium	mg kg-1 (ppm)

Mn	Extractable manganese	mg kg-1 (ppm)

N	Organic nitrogen	mg kg-1 (ppm)

Na	Extractable sodium	mg kg-1 (ppm)

P	Extractable phosphorus	mg (100 kg-1)

Ptot	Total phosphorus	mg (100 kg-1)

Zn	Extractable zinc	mg kg-1 (ppm)

References

Hengl T, Heuvelink GBM, Kempen B, Leenaars JGB, Walsh MG, Shepherd KD, et al. (2015) Mapping Soil Properties of Africa at 250 m Resolution: Random Forests Significantly Improve Current Predictions. PLoS ONE 10(6): e0125814. doi:10.1371/journal.pone.0125814

See also

Examples

# \donttest{
# downloads a large file
fe <- soil_af_elements("Fe", path=tempdir(), quiet=TRUE)
#> Cached as: /tmp/Rtmp0MapG8/soil_af/af_fe_0-30cm_30s.tif
# }