Access Geographic Data

Functions for downloading of geographic data for use in spatial analysis and mapping. The package facilitates access to climate, crops, elevation, land use, soil, species occurrence, accessibility, administrative boundaries and other data.

Function	Description

bio_oracle	Marine data from bio-oracle

cmip6_world	Downscaled and calibrated CMIP6 projected future climate data

cmip6_tile	Downscaled and calibrated CMIP6 data by tile

country_codes	Country codes

crop_calendar_sacks	Sachs crop calendar data

crop_monfreda	Monfreda crop data (area, yield)

crop_spam	SPAM crop data (area, yield, value)

cropland	Cropland density for the world from three sources

elevation_3s	Elevation data for tile (3 seconds resolution)

elevation_30s	Elevation data for by country (30 seconds resolution)

elevation_global	Global elevation data (various resolutions)

gadm	Administrative boundaries for any country in the world

world	Boundaries for the countries in the world

landcover	Global landcover data

footprint	Human footprint data

osm	OpenStreetMap data by country

population	Download population density data

soil_af	Chemical and physical soil properties data for Africa for different soil depths

soil_af_water	Physical soil properties for Africa for water balance computations

soil_af_elements	Soil element concentration data for Africa

soil_af_isda	Soil data for Africa derived from the iDSA data set

soil_world_vsi	Virtually connect to the global soilgrids data

soil_world	Global soils data

sp_occurrence	Species occurrence data from the Global Biodiversity Information Facility

travel_time	Travel time to cities and ports

worldclim_global	Global climate data

worldclim_country	Climate data by country

worldclim_tile	Climate data by tile