Soil data for Africa

Download chemical soil properties data for Africa for different soil depths. The spatial resolution is 30 arc-seconds (about 1 km2), aggregated from the original 250m resolution.

There are more recent estimations for some of the properties available in other data sets. See soil_af_isda and soil_world.

For more info, see https://web.archive.org/web/20250421195225/https://www.isric.org/projects/soil-property-maps-africa-250-m-resolution

The data have a CC-BY 4.0 NC license

soil_af(var, depth, path, ...)

Arguments

var

character. Variables name (see Details below). One of: "acid-exch", "Al-extr", "Al-exch", "bases-exch", "BDR", "BLKD", "Ca-exch", "CEC", "clay", "coarse", "drain", "ECN", "K-exch", "Mg-exch", "Na-exch", "Ntot", "pH", "sand", "silt", "SOC"

depth

numeric. One of 5, 15, 30, 60, 100, 200. This is shorthand for the following depth ranges: 0-5, 5-15, 15-30, 30-60, 60-100, 100-200 cm. Or one of 20, 50 for 0-20 or 20-50 cm

path

character. Path for storing the downloaded data. See geodata_path

...

additional arguments passed to download.file

Value

SpatRaster

Details

var	description	unit

clay	Soil texture fraction clay	%

sand	Soil texture fraction sand	%

silt	Soil texture fraction silt	%

coarse	Coarse fragments volumetric	%

SOC	Organic carbon	g kg-1 (‰)

BLKD	Bulk density (fine earth)	kg m-3

BDR	Depth to bedrock	cm

.	.	.

pH	pH (H2O)	-

ECN	Electrical conductivity	mS/m (?)

acid-exch	Exchangeable acidity	cmol(+) kg-1

bases-exch	Sum of exchangeable bases	cmol(+) kg-1

CEC	Cation Exchange Capacity	cmol(+) kg-1

Al-extr	Extractable Aluminum (Mehlich 3)	mg kg-1 (ppm)

Al-exch	Exchangeable Aluminum	cmol(+) kg-1

Ca-exch	Exchangeable Calcium	cmol(+) kg-1

K-exch	Exchangeable Potassium	cmol(+) kg-1

Mg-exch	Exchangeable Magnesium	cmol(+) kg-1

Na-exch	Exchangeable Sodium	cmol(+) kg-1

Ntot	Total nitrogen	g kg-1

References

Hengl T, Heuvelink GBM, Kempen B, Leenaars JGB, Walsh MG, Shepherd KD, et al. (2015) Mapping Soil Properties of Africa at 250 m Resolution: Random Forests Significantly Improve Current Predictions. PLoS ONE 10(6): e0125814. doi:10.1371/journal.pone.0125814

See also

Examples

# \donttest{
# downloads a large file
aph <- soil_af(var="ph", depth=5, path=tempdir())
#> Cached as: /tmp/RtmpYfsEyU/soil_af/af_ph_0-5cm_30s.tif
# }