iSDA soil data for Africa, virtual connection

Virtually connect to the iSDA soil data for Africa. The spatial of these data is 30m.

For more info see:

https://envirometrix.nl/isdasoil-open-soil-data-for-africa/

https://zenodo.org/search?page=1&size=20&q=iSDAsoil

Usage

soil_af_isda_vsi(var)

Arguments

var

character. The variables name, one of: "Al", "bdr", "clay", "C.tot", "CEC", "Ca", "db.od", "eCEC.f", "Fe", "K", "Mg", "N.tot", "oc", "P", "pH.H2O", "sand", "silt", "S", "texture", "wpg2", "Zn".see Details

Value

SpatRaster

Details

var	description	unit

Al	extractable aluminum	mg kg-1

bdr	bed rock depth	cm

clay	clay content	%

C.tot	total carbon	kg-1

Ca	extractable calcium	mg kg-1

db.od	bulk density	kg m-3

eCEC.f	effective cation exchange capacity	cmol(+) kg-1

Fe	extractable iron	mg kg-1

K	extractable potassium	mg kg-1

Mg	extractable magnesium	mg kg-1

N.tot	total organic nitrogen	g kg-1

OC	Organic Carbon	g kg-1

P	Phosphorus content	mg kg-1

pH.H2O	pH (H2O)	-

sand	Sand content	%

silt	Silt content	%

S	Extractable sulfer	mg kg-1

texture	texture class	-

wpg2	stone content	%

Zn	Extractable zinc	mg kg-1

References

Tomislav Hengl, Matthew A. E. Miller, Josip Križan, Keith D. Shepherd, Andrew Sila, Milan Kilibarda, Ognjen Antonijevic, Luka Glušica, Achim Dobermann, Stephan M. Haefele, Steve P. McGrath, Gifty E. Acquah, Jamie Collinson, Leandro Parente, Mohammadreza Sheykhmousa, Kazuki Saito, Jean-Martial Johnson, Jordan Chamberlin, Francis B.T. Silatsa, Martin Yemefack, John Wendt, Robert A. MacMillan, Ichsani Wheeler & Jonathan Crouch, 2021. African soil properties and nutrients mapped at 30 m spatial resolution using two-scale ensemble machine learning. Scientific Reports 11: 6130.

See also